Refactor (#780)
* reorganisation of server

* split up and simplified assets middleware

* further simplification
