+++ node/lib/handlebars.js
