+++ node/lib/verify-assets-exist.js
