+++ node/models/navigation/decorate.js
