+++ node/test/navigation/navigation.test.js
@@ -10,7 +10,7 @@ describe('Navigation middleware', () => {
 	before(() => {
 		const NavigationModelV1Stub = sinon.stub().returns(navigationModelV1Stub);
 		const NavigationModelV2Stub = sinon.stub().returns(navigationModelV2Stub);
-		navigation = proxyquire('../../src/navigation/index', {'./navigationModelV1':NavigationModelV1Stub, './navigationModelV2':NavigationModelV2Stub});
+		navigation = proxyquire('../../models/navigation/index', {'./navigationModelV1':NavigationModelV1Stub, './navigationModelV2':NavigationModelV2Stub});
 	});
 
 	afterEach(() => {
